+++ test/unit/CMakeLists.txt
@@ -2,8 +2,6 @@ add_scylla_test(bptree_compaction_test
   KIND UNIT)
 add_scylla_test(btree_compaction_test
   KIND UNIT)
-add_scylla_test(btree_stress_test
-  KIND UNIT)
 add_scylla_test(cross_shard_barrier_test
   KIND UNIT)
 add_scylla_test(lsa_sync_eviction_test
